scripts/analysis/requirements.txt
pandas
yfinance
rich
fpdf2
